output/rpa_envio_candidaturas/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chardet==5.2.0
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
lxml==5.2.2
openai==1.35.10
outcome==1.3.0.post0
packaging==24.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pandas==2.2.2
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pypng==0.20220715.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tqdm==4.66.4
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
webdriver-manager==4.0.1
webdrivermanager==0.10.0
websocket-client==1.8.0
wsproto==1.2.0